requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
langgraph
sentence-transformers
pypdf
chromadb
bs4
